requirements_render.txt
# Requirements for Render.com deployment (DEMO_MODE=true)
# Excludes: pyodbc, pyzk, pyinstaller, pywin32-ctypes (hardware/Windows-only)

fastapi==0.135.3
uvicorn==0.43.0
sqlalchemy==2.0.49
python-dotenv==1.2.2
python-multipart==0.0.24
python-jose==3.5.0
pydantic==2.12.5
pydantic-core==2.41.5
starlette==1.0.0
anyio==4.13.0
typing-extensions==4.15.0
typing-inspection==0.4.2

# Data processing
pandas==3.0.2
numpy==2.4.4
openpyxl==3.1.5
xlsxwriter==3.2.9
python-dateutil==2.9.0.post0

# HTTP / WebSocket
httpx==0.28.1
httpcore==1.0.9
websockets==16.0
requests==2.33.1

# Auth
passlib==1.7.4
bcrypt==3.2.0

# Other
pyyaml==6.0.3
six==1.17.0
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
urllib3==2.6.3
h11==0.16.0
click==8.3.2
annotated-types==0.7.0
annotated-doc==0.0.4
sniffio==1.3.1

# SQLite async support (no longer needs pyodbc MSSQL driver)
aiosqlite==0.21.0
